fixes for trilinos 11. EPetra now supports long and int which causes ambiguous calls when using unsigned int.
git-svn-id: https://svn.dealii.org/trunk@27170 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/include/deal.II/lac/trilinos_sparse_matrix.h
@@ -2775,7 +2775,7 @@ namespace TrilinosWrappers
                                    // add the possibility to insert new values,
                                    // and in the second we just replace
                                    // data.
-    if (row_partitioner().MyGID(row) == true)
+    if (row_partitioner().MyGID(static_cast<int>(row)) == true)
       {
         if (matrix->Filled() == false)
           {
@@ -2992,7 +2992,7 @@ namespace TrilinosWrappers
                                    // can directly call the Epetra_CrsMatrix
                                    // input function, which is much faster
                                    // than the Epetra_FECrsMatrix function.
-    if (row_partitioner().MyGID(row) == true)
+    if (row_partitioner().MyGID(static_cast<int>(row)) == true)
       {
         ierr = matrix->Epetra_CrsMatrix::SumIntoGlobalValues(row, n_columns,
                                          const_cast<double*>(col_value_ptr),
@@ -3031,7 +3031,7 @@ namespace TrilinosWrappers
         std::cout << std::endl << std::endl;
         std::cout << "Matrix row has the following indices:" << std::endl;
         int n_indices, *indices;
-        trilinos_sparsity_pattern().ExtractMyRowView(row_partitioner().LID(row),
+        trilinos_sparsity_pattern().ExtractMyRowView(row_partitioner().LID(static_cast<int>(row)),
                                                      n_indices,
                                                      indices);
         for (int i=0; i<n_indices; ++i)
